+++ src/hardware/mame/saa1099.cpp
@@ -178,7 +178,7 @@ void saa1099_device::device_start()
 	m_sample_rate = clock() / 256;
 
 	/* for each chip allocate one stream */
-	m_stream = stream_alloc(0, 2, m_sample_rate);
+	m_stream = stream_alloc(0, 2, (int)m_sample_rate);
 
 	save_item(NAME(m_noise_params));
 	save_item(NAME(m_env_enable));
